% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.r
\name{AhriTreClient}
\alias{AhriTreClient}
\title{Create AHRI TRE client}
\usage{
AhriTreClient(
  api = CApi(),
  runtime_config = RuntimeConfig(),
  check_compatibility = TRUE
)
}
\arguments{
\item{api}{C API object}

\item{runtime_config}{Runtime config}

\item{check_compatibility}{Check protocol compatibility}
}
\value{
AhriTreClient object
}
\description{
Create AHRI TRE client
}
