% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payloads.r
\name{arrow_ipc_to_table}
\alias{arrow_ipc_to_table}
\title{Convert Arrow IPC payload to table}
\usage{
arrow_ipc_to_table(payload)
}
\arguments{
\item{payload}{Payload object}
}
\value{
Arrow table
}
\description{
Convert Arrow IPC payload to table
}
