% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.r
\name{CApi}
\alias{CApi}
\title{Create C API object}
\usage{
CApi(artifact = NULL, library_path = NULL)
}
\arguments{
\item{artifact}{RuntimeArtifact object (optional)}

\item{library_path}{Path to library (optional)}
}
\value{
C API object
}
\description{
Create C API object
}
