% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.r
\name{close.ahri_tre_client}
\alias{close.ahri_tre_client}
\title{Close AHRI TRE client}
\usage{
\method{close}{ahri_tre_client}(con, ...)
}
\arguments{
\item{con}{Client to close}

\item{...}{Additional arguments}
}
\description{
Close AHRI TRE client
}
