% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.r
\name{execute_json}
\alias{execute_json}
\title{Execute JSON protocol request}
\usage{
execute_json(client, request)
}
\arguments{
\item{client}{AhriTreClient object}

\item{request}{Request object}
}
\value{
Protocol result
}
\description{
Execute JSON protocol request
}
