% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.r
\name{format_try_error}
\alias{format_try_error}
\title{Format a try-error for display}
\usage{
format_try_error(err)
}
\arguments{
\item{err}{Error object}
}
\value{
Formatted error message
}
\description{
Format a try-error for display
}
