% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.r
\name{is_tre_error}
\alias{is_tre_error}
\title{Check if error is a known TRE error type}
\usage{
is_tre_error(err, type = NULL)
}
\arguments{
\item{err}{Error object}

\item{type}{Error type to check}
}
\value{
Logical
}
\description{
Check if error is a known TRE error type
}
