% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payloads.r
\name{Payload}
\alias{Payload}
\title{Create a payload object}
\usage{
Payload(
  kind,
  protocol_ref = NULL,
  media_type = NULL,
  suggested_name = NULL,
  size_bytes = 0,
  data = NULL
)
}
\arguments{
\item{kind}{Payload kind}

\item{protocol_ref}{Protocol reference}

\item{media_type}{Media type}

\item{suggested_name}{Suggested file name}

\item{size_bytes}{Size in bytes}

\item{data}{Raw data}
}
\value{
Payload object
}
\description{
Create a payload object
}
