% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.r
\name{protocol_failure_summary}
\alias{protocol_failure_summary}
\title{Summarize protocol failure}
\usage{
protocol_failure_summary(envelope)
}
\arguments{
\item{envelope}{Response envelope}
}
\value{
Failure summary
}
\description{
Summarize protocol failure
}
