% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.r
\name{runtime_c_abi_library}
\alias{runtime_c_abi_library}
\title{Get path to the C ABI shared library}
\usage{
runtime_c_abi_library(artifact)
}
\arguments{
\item{artifact}{An \code{ahri_tre_runtime_artifact} object.}
}
\value{
Character path to the shared library.
}
\description{
Determines the correct library name and extension based on the OS.
Override can be specified in the manifest under \code{artifacts.c_abi_library}.
}
