% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.r
\name{runtime_c_header}
\alias{runtime_c_header}
\title{Get path to the C header file}
\usage{
runtime_c_header(artifact)
}
\arguments{
\item{artifact}{An \code{ahri_tre_runtime_artifact} object.}
}
\value{
Character path to \code{ahri_tre_ffi_c.h}.
}
\description{
Get path to the C header file
}
