% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.r
\name{runtime_daemon_binary}
\alias{runtime_daemon_binary}
\title{Get path to the daemon binary}
\usage{
runtime_daemon_binary(artifact)
}
\arguments{
\item{artifact}{An \code{ahri_tre_runtime_artifact} object.}
}
\value{
Character path to the daemon executable.
}
\description{
Adds \code{.exe} suffix on Windows.
}
