% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.r
\name{runtime_discover_daemon_binary}
\alias{runtime_discover_daemon_binary}
\title{Discover daemon binary}
\usage{
runtime_discover_daemon_binary(api = CApi(), config = RuntimeConfig())
}
\arguments{
\item{api}{C API object}

\item{config}{Runtime config}
}
\value{
Discovery result
}
\description{
Discover daemon binary
}
