% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.r
\name{runtime_ensure}
\alias{runtime_ensure}
\title{Ensure runtime is running}
\usage{
runtime_ensure(api = CApi(), config = RuntimeConfig())
}
\arguments{
\item{api}{C API object}

\item{config}{Runtime config}
}
\value{
Ensure result
}
\description{
Ensure runtime is running
}
