% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.r
\docType{data}
\name{RUNTIME_ROOT_ENV}
\alias{RUNTIME_ROOT_ENV}
\title{Environment variable used to specify the runtime root path}
\format{
An object of class \code{character} of length 1.
}
\usage{
RUNTIME_ROOT_ENV
}
\description{
Environment variable used to specify the runtime root path
}
\keyword{datasets}
