% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.r
\name{runtime_status}
\alias{runtime_status}
\title{Get runtime status}
\usage{
runtime_status(api = CApi(), config = RuntimeConfig())
}
\arguments{
\item{api}{C API object}

\item{config}{Runtime config}
}
\value{
Runtime status
}
\description{
Get runtime status
}
