% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.r
\name{runtime_stop}
\alias{runtime_stop}
\title{Stop runtime}
\usage{
runtime_stop(api = CApi(), config = RuntimeConfig())
}
\arguments{
\item{api}{C API object}

\item{config}{Runtime config}
}
\value{
Stop result
}
\description{
Stop runtime
}
