% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.r
\name{RuntimeConfig}
\alias{RuntimeConfig}
\title{Create runtime configuration}
\usage{
RuntimeConfig(
  daemon_endpoint = NULL,
  daemon_binary = NULL,
  readiness_timeout_ms = 0L,
  never_start = FALSE
)
}
\arguments{
\item{daemon_endpoint}{Daemon endpoint URL}

\item{daemon_binary}{Daemon binary path}

\item{readiness_timeout_ms}{Timeout in milliseconds}

\item{never_start}{Never start daemon}
}
\value{
Runtime config object
}
\description{
Create runtime configuration
}
