% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.r
\docType{data}
\name{SUPPORTED_PROTOCOL_MIN}
\alias{SUPPORTED_PROTOCOL_MIN}
\title{Supported protocol version range for this R package}
\format{
An object of class \code{character} of length 1.
}
\usage{
SUPPORTED_PROTOCOL_MIN
}
\description{
Supported protocol version range for this R package
}
\keyword{datasets}
