\name{tre-command-wrappers}
\alias{asset_delete}
\alias{asset_duo_clear}
\alias{asset_duo_list}
\alias{asset_duo_replace}
\alias{asset_get}
\alias{asset_list}
\alias{asset_versions}
\alias{datafile_delete}
\alias{datafile_export}
\alias{datafile_list}
\alias{datafile_metadata}
\alias{datafile_search}
\alias{dataset_data}
\alias{dataset_delete}
\alias{dataset_export}
\alias{dataset_list}
\alias{dataset_metadata}
\alias{dataset_preview}
\alias{dataset_search}
\alias{dataset_withdraw}
\alias{auth_login}
\alias{auth_logout}
\alias{auth_status}
\alias{daemon_doctor}
\alias{daemon_start}
\alias{daemon_status}
\alias{daemon_stop}
\alias{daemon_version}
\alias{session_close}
\alias{session_list}
\alias{session_open}
\alias{session_reopen}
\alias{session_status}
\alias{session_use}
\alias{datastore_adopt}
\alias{datastore_create}
\alias{datastore_info}
\alias{datastore_list}
\alias{datastore_ping}
\alias{datastore_rotate}
\alias{datastore_schema}
\alias{domain_add}
\alias{domain_delete}
\alias{domain_get}
\alias{domain_list}
\alias{tag_get}
\alias{tag_list}
\alias{tag_set}
\alias{variable_add}
\alias{variable_delete}
\alias{variable_get}
\alias{variable_list}
\alias{variable_search}
\alias{variable_update}
\alias{vocabulary_add}
\alias{vocabulary_delete}
\alias{vocabulary_get}
\alias{vocabulary_list}
\alias{entity_delete}
\alias{entity_instance_asset_link_add}
\alias{entity_instance_asset_link_list}
\alias{entity_instance_dataset_link_add}
\alias{entity_instance_dataset_link_get}
\alias{entity_instance_dataset_link_list}
\alias{entity_instance_datasets}
\alias{entity_instance_ensure}
\alias{entity_instance_add}
\alias{entity_instance_get}
\alias{entity_instance_list}
\alias{entity_instance_map_add}
\alias{entity_instance_map_get}
\alias{entity_instance_map_list}
\alias{entity_add}
\alias{entity_get}
\alias{entity_list}
\alias{entity_search}
\alias{entity_relation_delete}
\alias{entity_relation_instance_asset_link_add}
\alias{entity_relation_instance_asset_link_list}
\alias{entity_relation_instance_dataset_link_add}
\alias{entity_relation_instance_dataset_link_get}
\alias{entity_relation_instance_dataset_link_list}
\alias{entity_relation_instance_ensure}
\alias{entity_relation_instance_add}
\alias{entity_relation_instance_get}
\alias{entity_relation_instance_list}
\alias{entity_relation_instance_map_add}
\alias{entity_relation_instance_map_get}
\alias{entity_relation_instance_map_list}
\alias{entity_relation_add}
\alias{entity_relation_get}
\alias{entity_relation_list}
\alias{entity_relation_search}
\alias{ingest_datafile}
\alias{ingest_dataset_datafile}
\alias{ingest_dataset_sql}
\alias{ingest_dataset_table}
\alias{ingest_redcap_project}
\alias{transformation_list}
\alias{completion}
\alias{doctor}
\alias{schema_get}
\alias{schema_list}
\alias{version}
\alias{study_access_grant}
\alias{study_access_list}
\alias{study_access_revoke}
\alias{study_add}
\alias{study_add_domain}
\alias{study_clear_current}
\alias{study_context_list}
\alias{study_current}
\alias{study_custodians_add}
\alias{study_custodians_list}
\alias{study_custodians_remove}
\alias{study_custodians_transfer}
\alias{study_delete}
\alias{study_duo_list}
\alias{study_duo_replace}
\alias{study_get}
\alias{study_list}
\alias{study_search}
\alias{study_use}
\title{Generated TRE Command Wrapper Functions}
\description{Auto-generated wrappers for TRE protocol commands.}
\details{See package documentation.}
\keyword{package}
